se documenta el codigo

+++ src/main/java/com/example/todoList/services/TareasServices.java
@@ -1,4 +1,3 @@
-
 package com.example.todoList.services;
 
 import com.example.todoList.model.Message;
@@ -13,76 +12,80 @@ import org.springframework.stereotype.Service;
  *
  * @author Nico Morales
  */
-
 @Service
-public class TareasServices  implements ITareasServices{
-    
+public class TareasServices implements ITareasServices {
+
     @Autowired
     private ITareaRepository tareaRepo;
 
     @Override
-    public List <Tareas> traerTareas() {
-      
-       return tareaRepo.findAll();
+    public List<Tareas> traerTareas() {
+
+        return tareaRepo.findAll();
     }
 
     @Override
     public Optional<Tareas> traerTarea(Long id) {
-        
-        Tareas tarea = new Tareas();
+
         return tareaRepo.findById(id);
     }
 
     @Override
     public Message crearTareas(Tareas tarea) {
-       
+
         try {
-            
+
             if (tareaRepo.findByTitle(tarea.getTitulo()).isEmpty()) {
-                
-               tareaRepo.save(tarea);
-               return new Message("Tarea creada con Exito!!"); 
-            
+
+                tareaRepo.save(tarea);
+                return new Message("Tarea creada con Exito!!");
+
             } else {
-                
-                return new Message("La Tarea ya Existe!!"); 
+
+                return new Message("La Tarea ya Existe!!");
 
             }
-           
-            
+
         } catch (Exception e) {
-            
-            return new Message("Error, "+e);
+
+            return new Message("Error, " + e);
         }
     }
 
     @Override
     public Message borrarTarea(Long id) {
-      
+
         try {
-            
-            tareaRepo.deleteById(id);
-            return new Message("Tarea borrada con Exito!!!");
-            
+
+            if (this.traerTarea(id).isEmpty()) {
+
+                return new Message("La Tarea, No Existe!!!");
+
+            } else {
+
+                tareaRepo.deleteById(id);
+                return new Message("Tarea borrada con Exito!!!");
+            }
+
         } catch (Exception e) {
-            
-            return new Message("Error, "+e);
+
+            return new Message("Error, " + e);
         }
     }
 
     @Override
     public Message editarTarea(Tareas tarea) {
-        
-          try {
-            
+
+        try {
+
             tareaRepo.save(tarea);
             return new Message("Tarea modificada con Exito!!");
-            
+
         } catch (Exception e) {
-            
-            return new Message("Error, "+e);
+
+            return new Message("Error, " + e);
 
         }
     }
-    
+
 }
